requirements.txt
aiogram==2.25.1
python-dotenv==1.0.0
sqlalchemy==2.0.11
lms-synergy-library==0.2.2
